Product Recco/requirements.txt
streamlit
openai==0.28.1
requests
Pillow